# chatgpt-gpt-openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pivota Shopping Assistant API",
    "version": "2.0.0",
    "description": "API for shopping operations including product search, quote-first safe checkout, payment processing, and order management. Checkout v2 requires quote_id for create_order, idempotency_key for money writes, confirmation_token for submit_payment, and treats expected_amount as a verification echo only."
  },
  "servers": [
    {
      "url": "https://pivota-agent-production.up.railway.app"
    }
  ],
  "paths": {
    "/agent/shop/v1/invoke": {
      "post": {
        "summary": "Execute shopping operation",
        "description": "Unified endpoint for all shopping operations",
        "operationId": "invokeShoppingOperation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShoppingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ShoppingRequest": {
        "type": "object",
        "required": ["operation", "payload"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": [
              "find_products",
              "find_products_multi",
              "get_product_detail",
              "preview_quote",
              "create_order",
              "submit_payment",
              "get_order_status",
              "request_after_sales"
            ],
            "description": "The shopping operation to perform. Use find_products_multi when you want to search across all merchants in Pivota Infra."
          },
          "payload": {
            "type": "object",
            "description": "Operation-specific payload",
            "properties": {
              "acp_state": {
                "type": "object",
                "description": "Opaque ACP state from previous calls"
              },
              "ap2_state": {
                "type": "object",
                "description": "Opaque AP2 payment state"
              },
              "idempotency_key": {
                "type": "string",
                "minLength": 8,
                "description": "Required for create_order and submit_payment. Client-generated stable key for write dedupe."
              },
              "confirmation_token": {
                "type": "string",
                "description": "Required for submit_payment. Server-minted proof that the user confirmed the quote UI."
              },
              "search": {
                "type": "object",
                "description": "Product search parameters. If merchant_id is omitted, the backend will search across all available merchants that the current Pivota environment exposes.",
                "properties": {
                  "merchant_id": {
                    "type": "string",
                    "description": "Optional merchant ID filter. Omit to let Pivota automatically route across all merchants."
                  },
                  "query": {
                    "type": "string",
                    "description": "Search query"
                  },
                  "price_min": {
                    "type": "number",
                    "description": "Minimum price"
                  },
                  "price_max": {
                    "type": "number",
                    "description": "Maximum price"
                  },
                  "category": {
                    "type": "string",
                    "description": "Product category"
                  },
                  "in_stock_only": {
                    "type": "boolean",
                    "description": "Filter for in-stock items",
                    "default": true
                  },
                  "page": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "page_size": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                  }
                }
              },
              "product": {
                "type": "object",
                "description": "Product detail parameters",
                "required": ["merchant_id", "product_id"],
                "additionalProperties": true,
                "properties": {
                  "merchant_id": {
                    "type": "string",
                    "description": "Merchant ID"
                  },
                  "product_id": {
                    "type": "string",
                    "description": "Product ID"
                  }
                }
              },
              "order": {
                "type": "object",
                "description": "Order creation parameters. create_order must be called with payload.idempotency_key and a locked quote_id. Caller-supplied items, prices, and amounts are not accepted.",
                "required": ["quote_id", "shipping_address"],
                "additionalProperties": false,
                "properties": {
                  "quote_id": {
                    "type": "string",
                    "description": "Required. Server-issued quote id from preview_quote; all pricing is copied from the locked quote."
                  },
                  "shipping_address": {
                    "type": "object",
                    "required": ["country", "city", "address_line1", "recipient_name"],
                    "additionalProperties": false,
                    "properties": {
                      "recipient_name": {
                        "type": "string"
                      },
                      "address_line1": {
                        "type": "string"
                      },
                      "address_line2": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      },
                      "country": {
                        "type": "string"
                      },
                      "postal_code": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      }
                    }
                  },
                  "delivery_preferences": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "max_eta_days": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "preferred_carriers": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              },
              "payment": {
                "type": "object",
                "description": "Payment submission parameters. submit_payment requires payload.idempotency_key and payload.confirmation_token. expected_amount is a verification echo; the charged amount is read server-side from the order/quote snapshot.",
                "required": ["order_id", "expected_amount", "currency"],
                "additionalProperties": false,
                "properties": {
                  "order_id": {
                    "type": "string"
                  },
                  "expected_amount": {
                    "type": "number",
                    "description": "Echo of the amount the user confirmed, for verification only. A mismatch with the server-side order/quote snapshot hard-fails."
                  },
                  "currency": {
                    "type": "string"
                  },
                  "payment_method_hint": {
                    "type": "string",
                    "enum": ["card", "wallet", "bank_transfer", "pay_later", "shop_pay"]
                  },
                  "payment_handler_id": {
                    "type": "string"
                  },
                  "payment_handler_type": {
                    "type": "string"
                  },
                  "return_url": {
                    "type": "string"
                  }
                }
              },
              "status": {
                "type": "object",
                "description": "Order status/after-sales parameters",
                "properties": {
                  "order_id": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "requested_action": {
                    "type": "string",
                    "enum": ["refund", "return", "exchange", "support"]
                  }
                }
              }
            }
          }
        }
      },
      "ShoppingResponse": {
        "type": "object",
        "description": "Response varies by operation type"
      }
    }
  }
}